chore: revert site to static storefront
