--- .env.example
@@ -1,2 +1,0 @@
-# Shared local development env vars
-NEXT_PUBLIC_API_BASE=http://localhost:8787
