--- .github/workflows/deploy-pages.yml
@@ -1,40 +1,0 @@
-name: Deploy Pages
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - 'apps/web/**'
-      - '.github/workflows/deploy-pages.yml'
-      - 'package.json'
-      - 'pnpm-lock.yaml'
-      - 'pnpm-workspace.yaml'
-
-jobs:
-  deploy:
-    runs-on: ubuntu-latest
-    env:
-      NEXT_PUBLIC_API_BASE: https://api.diary.liuallen.com
-    steps:
-      - uses: actions/checkout@v4
-      - uses: pnpm/action-setup@v3
-        with:
-          version: 8
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-          cache: 'pnpm'
-      - run: pnpm install --frozen-lockfile
-      - name: Build static site
-        run: |
-          pnpm --filter web build
-          pnpm --filter web export
-      - name: Publish to Cloudflare Pages
-        uses: cloudflare/pages-action@v1
-        with:
-          apiToken: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-          accountId: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-          projectName: maildiary-web
-          directory: apps/web/out
-          gitHubToken: ${{ secrets.GITHUB_TOKEN }}
-          branch: main
