--- .github/workflows/deploy-worker.yml
@@ -1,34 +1,0 @@
-name: Deploy Worker
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - 'apps/worker/**'
-      - 'wrangler.toml'
-      - 'migrations/**'
-      - '.github/workflows/deploy-worker.yml'
-      - 'package.json'
-      - 'pnpm-lock.yaml'
-      - 'pnpm-workspace.yaml'
-
-jobs:
-  deploy:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: pnpm/action-setup@v3
-        with:
-          version: 8
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-          cache: 'pnpm'
-      - run: pnpm install --frozen-lockfile
-      - name: Typecheck Worker
-        run: pnpm --filter worker typecheck
-      - name: Deploy Worker
-        run: pnpm --filter worker run deploy
-        env:
-          CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-          CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
