+++ .gitignore
@@ -1,6 +1,61 @@
-node_modules
-.pnpm-store
+# 系统文件
 .DS_Store
+.DS_Store?
+._*
+.Spotlight-V100
+.Trashes
+ehthumbs.db
+Thumbs.db
+Desktop.ini
+
+# 编辑器和IDE
+.vscode/
+.idea/
+*.swp
+*.swo
+*~
+.project
+.settings/
+*.sublime-project
+*.sublime-workspace
+
+# 日志文件
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+logs/
+
+# 临时文件
+.tmp/
+.cache/
+*.tmp
+*.bak
+
+# 备份文件
+*.backup
+*.old
+
+# Node modules (如果将来使用)
+node_modules/
+package-lock.json
+yarn.lock
+
+# 构建输出 (如果将来使用构建工具)
+dist/
+build/
+*.min.js
+*.min.css
+
+# 环境变量
+.env
 .env.local
-apps/web/.next
-apps/web/out
+.env.*.local
+
+# 测试覆盖率
+coverage/
+.nyc_output/
+
+# 操作系统
+*.swp
+.directory
