+++ apps/hanzi-teacher/js/app.js
