+++ apps/hanzi-teacher/js/speak.js
