--- apps/web/next.config.js
@@ -1,10 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  output: 'export',
-  images: {
-    unoptimized: true,
-  },
-};
-
-module.exports = nextConfig;
