--- apps/web/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "web",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "typecheck": "tsc --noEmit",
-    "export": "next export"
-  },
-  "dependencies": {
-    "next": "14.2.3",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "swr": "2.2.5"
-  },
-  "devDependencies": {
-    "@types/node": "20.12.7",
-    "@types/react": "18.2.21",
-    "@types/react-dom": "18.2.7",
-    "typescript": "5.4.5"
-  }
-}
