--- apps/web/styles/globals.css
@@ -1,63 +1,0 @@
-:root {
-  font-family: 'Inter', system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  color: #0f172a;
-  background-color: #f8fafc;
-}
-
-body {
-  margin: 0;
-  background-color: #f8fafc;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-main {
-  padding: 2rem;
-  max-width: 960px;
-  margin: 0 auto;
-}
-
-button {
-  cursor: pointer;
-  border: none;
-  border-radius: 8px;
-  padding: 0.75rem 1.25rem;
-  background: #2563eb;
-  color: white;
-  font-weight: 600;
-}
-
-input,
-select,
-textarea {
-  border-radius: 8px;
-  border: 1px solid #cbd5f5;
-  padding: 0.75rem 1rem;
-  width: 100%;
-  font-size: 1rem;
-}
-
-.card {
-  background: white;
-  border-radius: 12px;
-  padding: 1.5rem;
-  box-shadow: 0 10px 30px rgba(15, 23, 42, 0.08);
-  margin-bottom: 1.5rem;
-}
-
-.flex {
-  display: flex;
-}
-
-.flex-between {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
