--- apps/worker/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "lib": ["ES2021"],
-    "types": ["@cloudflare/workers-types"],
-    "module": "esnext",
-    "target": "es2021",
-    "moduleResolution": "bundler",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "allowSyntheticDefaultImports": true,
-    "baseUrl": "./src"
-  },
-  "include": ["src/**/*"]
-}
