+++ assets/css/store.css
@@ -37,18 +37,6 @@ body {
     flex-direction: column;
 }
 
-.visually-hidden {
-    position: absolute;
-    width: 1px;
-    height: 1px;
-    padding: 0;
-    margin: -1px;
-    overflow: hidden;
-    clip: rect(0, 0, 0, 0);
-    white-space: nowrap;
-    border: 0;
-}
-
 .container {
     max-width: 1200px;
     margin: 0 auto;
@@ -112,8 +100,6 @@ body {
 .main {
     flex: 1;
     padding: 40px 0;
-    max-width: none;
-    margin: 0;
 }
 
 /* Hero区域 */
@@ -140,23 +126,6 @@ body {
     margin: 0 auto;
 }
 
-.hero-cta {
-    margin-top: 28px;
-    display: flex;
-    justify-content: center;
-}
-
-.hero-pill {
-    display: inline-flex;
-    align-items: center;
-    gap: 12px;
-    background: rgba(255, 255, 255, 0.12);
-    border-radius: 999px;
-    padding: 12px 24px;
-    font-size: 0.95rem;
-    letter-spacing: 0.03em;
-}
-
 /* Section标题 */
 .section-title {
     font-size: 2rem;
@@ -165,118 +134,9 @@ body {
     color: var(--text-primary);
 }
 
-.section-header {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    margin-bottom: 24px;
-}
-
-.section-subtitle {
-    color: var(--text-secondary);
-    font-size: 0.95rem;
-}
-
 /* 应用网格 */
 .apps-section {
     margin-bottom: 80px;
-}
-
-.filters {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 12px;
-    align-items: center;
-    margin-bottom: 16px;
-}
-
-.search-field {
-    position: relative;
-    flex: 1 1 280px;
-    min-width: 240px;
-}
-
-.search-field input {
-    width: 100%;
-    padding: 12px 16px 12px 44px;
-    border-radius: 999px;
-    border: 1px solid var(--border-color);
-    background: var(--bg-primary);
-    color: var(--text-primary);
-    font-size: 1rem;
-    transition: var(--transition);
-    box-shadow: inset 0 1px 2px rgba(0, 0, 0, 0.04);
-}
-
-.search-field input::placeholder {
-    color: var(--text-secondary);
-}
-
-.search-field input:focus {
-    outline: none;
-    border-color: rgba(0, 122, 255, 0.45);
-    box-shadow: 0 0 0 4px rgba(0, 122, 255, 0.12);
-}
-
-.search-icon {
-    position: absolute;
-    left: 18px;
-    top: 50%;
-    transform: translateY(-50%);
-    pointer-events: none;
-    font-size: 1.1rem;
-    color: var(--text-secondary);
-}
-
-.tag-filter-bar {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 10px;
-    margin-bottom: 24px;
-}
-
-.filter-placeholder {
-    color: var(--text-secondary);
-    font-size: 0.9rem;
-}
-
-.filter-chip {
-    padding: 8px 16px;
-    border-radius: 999px;
-    background: var(--bg-primary);
-    border: 1px solid rgba(0, 0, 0, 0.04);
-    color: var(--text-secondary);
-    font-size: 0.85rem;
-    font-weight: 600;
-    cursor: pointer;
-    transition: var(--transition);
-}
-
-.filter-chip:hover {
-    transform: translateY(-1px);
-    box-shadow: var(--shadow-sm);
-}
-
-.filter-chip.active {
-    background: var(--primary-color);
-    color: white;
-    box-shadow: var(--shadow-sm);
-}
-
-.btn-tertiary {
-    background: var(--bg-primary);
-    color: var(--primary-color);
-    border: 1px solid rgba(0, 122, 255, 0.25);
-}
-
-.btn-tertiary:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-    box-shadow: none;
-}
-
-.btn-tertiary:not(:disabled):hover {
-    background: rgba(0, 122, 255, 0.08);
 }
 
 .apps-grid {
@@ -284,22 +144,6 @@ body {
     grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
     gap: 24px;
     animation: fadeInUp 0.6s ease-out;
-}
-
-.empty-state {
-    grid-column: 1 / -1;
-    background: var(--bg-primary);
-    border-radius: var(--radius-md);
-    padding: 48px 24px;
-    text-align: center;
-    border: 1px dashed var(--border-color);
-    color: var(--text-secondary);
-}
-
-.empty-state h4 {
-    font-size: 1.3rem;
-    margin-bottom: 12px;
-    color: var(--text-primary);
 }
 
 @keyframes fadeInUp {
@@ -329,11 +173,6 @@ body {
     box-shadow: var(--shadow-md);
 }
 
-.app-card:focus-visible {
-    outline: 3px solid rgba(0, 122, 255, 0.35);
-    outline-offset: 4px;
-}
-
 .app-icon {
     font-size: 4rem;
     margin-bottom: 16px;
@@ -352,30 +191,6 @@ body {
     font-size: 0.95rem;
     margin-bottom: 16px;
     line-height: 1.5;
-}
-
-.app-meta {
-    display: flex;
-    align-items: center;
-    gap: 12px;
-    margin-bottom: 16px;
-    font-size: 0.85rem;
-}
-
-.app-version {
-    background: var(--bg-secondary);
-    border-radius: 999px;
-    padding: 4px 10px;
-    color: var(--text-secondary);
-    font-weight: 600;
-}
-
-.app-badge {
-    background: rgba(255, 149, 0, 0.15);
-    color: var(--warning-color);
-    border-radius: 999px;
-    padding: 4px 10px;
-    font-weight: 600;
 }
 
 .app-tags {
@@ -522,8 +337,6 @@ body {
     align-items: center;
     justify-content: center;
     border-radius: 50%;
-    background: transparent;
-    border: none;
 }
 
 .modal-close:hover {
@@ -623,15 +436,6 @@ body {
     .header .container {
         flex-direction: column;
         gap: 16px;
-    }
-
-    .filters {
-        flex-direction: column;
-        align-items: stretch;
-    }
-
-    .btn-tertiary {
-        width: 100%;
     }
 
     .section-title {
