+++ index.html
@@ -1,0 +1,93 @@
+<!DOCTYPE html>
+<html lang="zh-CN">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>刘Allen的应用商店</title>
+    <link rel="stylesheet" href="assets/css/store.css">
+</head>
+<body>
+    <!-- 顶部导航栏 -->
+    <header class="header">
+        <div class="container">
+            <h1 class="logo">📱 刘Allen的应用</h1>
+            <nav class="nav">
+                <a href="#apps" class="nav-link active">应用</a>
+                <a href="#about" class="nav-link">关于</a>
+            </nav>
+        </div>
+    </header>
+
+    <!-- 主内容区 -->
+    <main class="main">
+        <div class="container">
+            <!-- 欢迎区域 -->
+            <section class="hero">
+                <h2 class="hero-title">欢迎来到我的应用商店</h2>
+                <p class="hero-subtitle">探索有趣实用的Web应用</p>
+            </section>
+
+            <!-- 应用列表 -->
+            <section id="apps" class="apps-section">
+                <h3 class="section-title">精选应用</h3>
+
+                <div class="apps-grid" id="appsGrid">
+                    <!-- 应用卡片将由JavaScript动态生成 -->
+                </div>
+            </section>
+
+            <!-- 关于区域 -->
+            <section id="about" class="about-section">
+                <h3 class="section-title">关于</h3>
+                <div class="about-content">
+                    <p>这是一个个人应用商店，收集了我开发的各种有趣的Web应用。</p>
+                    <p>所有应用都是开源的，基于现代Web技术构建。</p>
+                    <div class="stats">
+                        <div class="stat-item">
+                            <div class="stat-number" id="appCount">0</div>
+                            <div class="stat-label">应用数量</div>
+                        </div>
+                        <div class="stat-item">
+                            <div class="stat-number">100%</div>
+                            <div class="stat-label">开源</div>
+                        </div>
+                        <div class="stat-item">
+                            <div class="stat-number">0</div>
+                            <div class="stat-label">广告</div>
+                        </div>
+                    </div>
+                </div>
+            </section>
+        </div>
+    </main>
+
+    <!-- 底部 -->
+    <footer class="footer">
+        <div class="container">
+            <p>&copy; 2025 刘Allen. All rights reserved.</p>
+            <p>
+                <a href="https://github.com/capmapt/liuallen" target="_blank" class="footer-link">GitHub</a>
+            </p>
+        </div>
+    </footer>
+
+    <!-- 应用模态框 -->
+    <div id="appModal" class="modal">
+        <div class="modal-content">
+            <span class="modal-close">&times;</span>
+            <div class="modal-body">
+                <div class="modal-icon" id="modalIcon"></div>
+                <h2 class="modal-title" id="modalTitle"></h2>
+                <p class="modal-description" id="modalDescription"></p>
+                <div class="modal-tags" id="modalTags"></div>
+                <div class="modal-actions">
+                    <button class="btn btn-primary" id="launchBtn">启动应用</button>
+                    <a class="btn btn-secondary" id="githubBtn" target="_blank">查看源码</a>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <script type="module" src="assets/js/store.js"></script>
+</body>
+</html>
