--- wrangler.toml
@@ -1,33 +1,0 @@
-name = "maildiary-worker"
-main = "apps/worker/src/index.ts"
-compatibility_date = "2024-05-01"
-workers_dev = false
-
-[vars]
-APP_BASE_URL = "https://diary.liuallen.com"
-API_BASE_URL = "https://api.diary.liuallen.com"
-COOKIE_DOMAIN = ".liuallen.com"
-MAIL_FROM = "reminder@diary.liuallen.com"
-
-[[d1_databases]]
-binding = "DB"
-database_name = "maildiary"
-database_id = "TODO_D1_DATABASE_ID"
-
-[[r2_buckets]]
-binding = "R2_BUCKET"
-bucket_name = "maildiary-attachments"
-
-[[kv_namespaces]]
-binding = "KV_SESSIONS"
-id = "TODO_KV_NAMESPACE_ID"
-
-[triggers]
-crons = ["*/5 * * * *"]
-
-[[routes]]
-pattern = "api.diary.liuallen.com/*"
-custom_domain = true
-
-[email]
-allowed_senders = ["reminder@diary.liuallen.com", "reply@diary.liuallen.com"]
